+++ src/player.h
@@ -2,12 +2,12 @@
 #define PLAYER_H
 
 #include <cstdlib>
-#include <iostream>
 #include <fstream>
+#include <iostream>
 #include <string>
 
-#include "nlohmann/json.hpp"
 #include "bee.h"
+#include "nlohmann/json.hpp"
 
 class Player {
     const std::string name;
@@ -16,9 +16,9 @@ class Player {
     uint32_t backpack_capacity;
     uint32_t pollen;
 
-    uint8_t backpack_upgrades;
-    uint8_t collect_amount_upgrades;
-    uint8_t honey_per_pollen_upgrades;
+    uint16_t backpack_upgrades;
+    uint16_t collect_amount_upgrades;
+    uint16_t honey_per_pollen_upgrades;
 
     std::vector<Bee> bees;
 
@@ -29,16 +29,12 @@ class Player {
     friend std::ostream &operator<<(std::ostream &out, const Player &player);
 
 public:
-    Player(
-        std::string player_name,
-        uint32_t player_honey,
-        uint32_t player_backpack_capacity,
-        uint32_t player_pollen,
-        uint8_t player_backpack_upgrades,
-        uint8_t player_collect_amount_upgrades,
-        uint8_t player_honey_per_pollen_upgrades,
-        std::vector<Bee> player_bees
-    );
+    Player(std::string player_name, uint32_t player_honey,
+           uint32_t player_backpack_capacity, uint32_t player_pollen,
+           uint16_t player_backpack_upgrades,
+           uint16_t player_collect_amount_upgrades,
+           uint16_t player_honey_per_pollen_upgrades,
+           std::vector<Bee> player_bees);
 
     /// load_player_stats reads the player_stats.json file
     /// and returns a Player object with its appropriate data
